% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cblmr.R
\name{blm_choice}
\alias{blm_choice}
\title{Choice of Best Linear Model}
\usage{
blm_choice(dataframe, response, ..., exp.comb)
}
\arguments{
\item{dataframe}{only numerical values are required (explanatory and response variables)}

\item{response}{dependent variable}

\item{exp.comb}{number of expected combinations}
}
\value{
a tab grouping predictions and goodness of fit per model
}
\description{
This function allows you to choose the best model from among different "best" models.
Evaluating based on predictions and goodness of fit.
}
\details{
Evaluating based on predictions (PRESS : Prediction Sum of Squares) and
goodness of fit (AIC, BIC, R Squared, Adj. Squared).
}
\author{
Jean Marie Cimula
}

